Return from fade on "None"
Fixes #52 with a return when both fade params are `None`. Also updates
the default pause value to not throw math errors on `0`.

(Added another `venv` variation to the ..gitignore)

Note: the values in the doc are incorrect; the functionality doesn't
seem to be suppoted on the shim or this isn't working as expected.